requirements.txt
Django>=3.1.4,<4.0
djangorestframework>=3.12.1,<3.12.2
django-cors-headers==3.5.0
#mysqlclient>=2.0.2,<2.1.0
PyJWT>=1.7.1,<1.7.3
django-environ==0.4.5
psycopg2==2.8.6
